Moves saturation functions in characteristiccurves
