+++ src/tdycore.c
@@ -1,5 +1,4 @@
 #include <private/tdycoreimpl.h>
-#include <private/tdysaturationimpl.h>
 #include <private/tdycharacteristiccurvesimpl.h>
 #include <private/tdympfao3Dcoreimpl.h>
 #include <private/tdympfaoimpl.h>
